+++ openpyxl2/reader/tests/test_worksheet.py
@@ -170,7 +170,7 @@ def test_hidden_row(datadir, WorkSheetParser):
     with open("hidden_rows_cols.xml", "rb") as src:
         rows = iterparse(src, tag='{%s}row' % SHEET_MAIN_NS)
         for _, row in rows:
-            parser.parse_row_row(row)
+            parser.parse_row(row)
     assert 2 in ws.row_dimensions
     assert dict(ws.row_dimensions[2]) == {'hidden': '1'}
 
@@ -184,7 +184,7 @@ def test_styled_row(datadir, WorkSheetParser):
     with open("complex-styles-worksheet.xml", "rb") as src:
         rows = iterparse(src, tag='{%s}row' % SHEET_MAIN_NS)
         for _, row in rows:
-            parser.parse_row_row(row)
+            parser.parse_row(row)
     assert 23 in ws.row_dimensions
     rd = ws.row_dimensions[23]
     assert rd.style_id == 28
@@ -408,10 +408,10 @@ def test_sheet_views(WorkSheetParser, datadir):
     parser = WorkSheetParser
 
     with open("frozen_view_worksheet.xml") as src:
-        sheet = fromstring(src.read())
+        sheet = src.read()
 
-    element = sheet.find("{%s}sheetViews" % SHEET_MAIN_NS)
-    parser.parse_sheet_views(element)
+    parser.source = sheet
+    parser.parse()
     ws = parser.ws
     view = ws.sheet_view
 
@@ -483,7 +483,7 @@ def test_row_dimensions(WorkSheetParser):
     element = fromstring(src)
 
     parser = WorkSheetParser
-    parser.parse_row_row(element)
+    parser.parse_row(element)
 
     assert 2 not in parser.ws.row_dimensions
 
@@ -525,7 +525,7 @@ def test_cell_without_coordinates(WorkSheetParser, datadir):
 
     parser = WorkSheetParser
     parser.shared_strings = ["Whatever"] * 10
-    parser.parse_row_row(el)
+    parser.parse_row(el)
 
     assert parser.ws.max_row == 1
     assert parser.ws.max_column == 5