+++ openpyxl2/worksheet/views.py
@@ -1,10 +1,18 @@
 from __future__ import absolute_import
 # Copyright (c) 2010-2016 openpyxl
 
-from openpyxl2.descriptors import Bool, Integer, String, Set, Float, Typed, NoneSet, Sequence
+from openpyxl2.descriptors import (
+    Bool,
+    Integer,
+    String,
+    Set,
+    Float,
+    Typed,
+    NoneSet,
+    Sequence,
+)
+from openpyxl2.descriptors.excel import ExtensionList
 from openpyxl2.descriptors.serialisable import Serialisable
-
-from openpyxl2.compat import safe_string
 
 
 class Pane(Serialisable):
@@ -76,7 +84,7 @@ class SheetView(Serialisable):
         self,
         windowProtection=None,
         showFormulas=None,
-        showGridLines=True,
+        showGridLines=None,
         showRowColHeaders=None,
         showZeros=None,
         rightToLeft=None,
@@ -120,11 +128,20 @@ class SheetView(Serialisable):
             selection = (Selection(), )
         self.selection = selection
 
-    def __iter__(self):
 
-        for attr in self.__attrs__:
-            value = getattr(self, attr)
-            if attr == 'showGridLines' and value:
-                continue
-            if value is not None:
-                yield attr, safe_string(value)
+class SheetViewList(Serialisable):
+
+    tagname = "sheetViews"
+
+    sheetView = Sequence(expected_type=SheetView, )
+    extLst = Typed(expected_type=ExtensionList, allow_none=True)
+
+    __elements__ = ('sheetView',)
+
+    def __init__(self,
+                 sheetView=None,
+                 extLst=None,
+                ):
+        if sheetView is None:
+            sheetView = [SheetView()]
+        self.sheetView = sheetView