+++ doc/changes.rst
@@ -1,5 +1,5 @@
-2.4.0 (unreleased)
-==================
+2.4.0-a1 (unreleased)
+=====================
 
 
 Minor changes